Move ESEF 2021 specific modules into subdirectory
